fix: fix details new screens
- fix date picker
- fix forms to add and edit activity and quiz
- fix inputs quiz
- fix name component whit files
- fix order files
- fix views
