--- app/(home)/activity/new/page.tsx
@@ -1,11 +1,0 @@
-import TaskForm from "@/app/components/courses/activities/newActivity";
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Go-escuela | Agregar Tarea",
-  description: "",
-};
-
-export default function Task() {
-  return <TaskForm />;
-}
