+++ app/(home)/courses/[slug]/activity/new/page.tsx
@@ -1,0 +1,11 @@
+import NewActivity from "@/app/components/courses/activity/NewActivity";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Go-escuela | Agregar Tarea",
+  description: "",
+};
+
+export default function Page() {
+  return <NewActivity />;
+}
