+++ app/css/style.css
@@ -10,6 +10,15 @@
 	}
 }
 
+	/* DatepickerStyles.css */
+	.datepicker .datepicker-disabled-dates-text {
+		color: #d1d5db; /* Color para fechas desactivadas */
+	}
+	
+	.datepicker .datepicker-enabled-dates-text {
+		color: #10b981; /* Color para fechas habilitadas */
+	}
+
 @layer components {}
 
 @layer utilities {
